修复了自动清除备份文件后，无法继续备份的问题

+++ ContextMenuManager/Methods/BackupHelper.cs
@@ -188,8 +188,8 @@ namespace ContextMenuManager.Methods
                     }
 
                 }
-                // 如果设备目录为空，则删除该设备目录
-                if (Directory.GetFiles(deviceDir).Length == 0)
+                // 如果设备目录为空且不为本机目录，则删除该设备目录
+                if ((Directory.GetFiles(deviceDir).Length == 0) && (deviceDir != AppConfig.ComputerHostName))
                 {
                     Directory.Delete(deviceDir);
                 }
